requirements.txt
streamlit
crewai==0.11.0
langchain-openai==0.0.2
matplotlib==3.8.2
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0